moved all the winning evaluation into a separate class
